+++ src/server/graphql/rootQuery.js
@@ -1,16 +1,18 @@
 import {GraphQLObjectType} from 'graphql';
 import User from 'server/graphql/types/User';
 import {getUserId} from 'server/utils/authorization';
-import activeProOrgCount from 'server/graphql/queries/activeProOrgCount';
-import activeProUserCount from 'server/graphql/queries/activeProUserCount';
-import countTiersForUser from 'server/graphql/queries/countTiersForUser';
+import suCountTiersForUser from 'server/graphql/queries/suCountTiersForUser';
+import suOrgCount from 'server/graphql/queries/suOrgCount';
+import suProOrgInfo from 'server/graphql/queries/suProOrgInfo';
+import suUserCount from 'server/graphql/queries/suUserCount';
 
 export default new GraphQLObjectType({
   name: 'Query',
   fields: () => ({
-    activeProOrgCount,
-    activeProUserCount,
-    countTiersForUser,
+    suCountTiersForUser,
+    suUserCount,
+    suProOrgInfo,
+    suOrgCount,
     viewer: {
       type: User,
       resolve: (source, args, {authToken, dataLoader}) => {